feat : 파일 구조화
